refactor: made a separate file for types declaration
close #91
