+++ frontend/routes/dashboard.tsx
@@ -2,19 +2,7 @@ import { createFileRoute } from "@tanstack/react-router";
 import axios from "axios";
 import API_URL from "../../src/config/api";
 import Dashboard from "../src/pages/Dashboard";
-
-export interface Ticket {
-	idTicket: number;
-	title: string;
-	description: string;
-	image: string | null;
-	level: "bas" | "moyen" | "haut" | "urgent";
-	createdAt: string;
-	updatedAt: string;
-	idStatus: "Ouvert" | "En cours" | "Fermé" | "Résolu";
-	idUser: number;
-	username: string;
-}
+import type { Ticket } from "../src/utils/types";
 
 export const Route = createFileRoute("/dashboard")({
 	loader: async (): Promise<Ticket[]> => {
